+++ src/bonsai/bonsai/bim/module/model/__init__.py
@@ -33,6 +33,7 @@ from . import (
     handler,
     host_add_opening_gizmo,
     mep,
+    mep_bend_preview,
     opening,
     product,
     profile,
@@ -273,11 +274,11 @@ classes = (
     mep.MEPUnjoinPair,
     mep.SelectMEPPathMembers,
     mep.MEPJoinSegments,
-    mep.EnableBendPreview,
-    mep.FinishBendPreview,
-    mep.CancelBendPreview,
-    mep.EnableBendPreviewFromBend,
-    mep.GizmoBendPreview,
+    mep_bend_preview.EnableBendPreview,
+    mep_bend_preview.FinishBendPreview,
+    mep_bend_preview.CancelBendPreview,
+    mep_bend_preview.EnableBendPreviewFromBend,
+    mep_bend_preview.GizmoBendPreview,
     mep.EnableEditingPipeSegment,
     mep.FinishEditingPipeSegment,
     mep.CancelEditingPipeSegment,
